wip: Change beacon card style
